Add a view for undergrad terms by campus
Each campus will have a collection of terms. This view will be the
source of that term data.

Term begin/end date & names are stored in ps_term_tbl, but they are
stored at the institution level. We need data at the Campus level. So
this view duplicates UMNTC term data as UMNRO term data.

We also need to identify which term is current. This is currently used
to determine which Term is automatically selected when a student is
searching for course fees.