Add key to Terms view
Terms will be a resources served via json, which means they will need an
ID value. The actual primary key of a term is the composite of
Institution and STRM, but JSON and composite keys don't work well
together. I'm creating a unique integer ID by combining campus & strm ,
hashing it and casting it to an integer.

Without the cast, every ID will end in `.0`, which doesn't work.