Add view for course fees.
This view is a tricky one.

The `tfms_data` part of the query is mostly lifted from the existing One
Stop Course Fees code, only with the UMNTC-only limit removed, along
with some unnecessary grouping.

But TFMS stores courses in a way that is incompatible with this service.
It will have all fees for a course in a fiscal year in a single row:

| year_id | course | fee amount | fall | summer | spring |
| 1       | 123    | 800.00     | 1    | 0      | 1      |

But this service displays fees as part of a subject, which in turn
belongs to a specific term. So that one fee needs to be 2 rows in my
table:

| strm   | amount | course |
| spring | 800.00 | 123    |
| fall   | 800.00 | 123    |

So I take the TFMS data and then create 3 tables (Summer, Spring, Fall)
and create the correct strm values.

Unioning these 3 tables together gives me the data I expect