Restructure resources to Courses and Fees-per-section
Tackling two things here

First, the previous approach of having courses within Subjects didn't
work well for Jenn & Drupal. We agreed that listing all Courses for a
Campus/Term would be better.

I'm doing that by creating a new Courses view, which is 1-row per
Course (compared to 1-row per fee in CourseFees).

Nice benefit of this change: When looking at a Campus/Term we only show
courses that have fees. No more dead-ends of looking in a Subject only
to find that it contains no courses with fees.

Second change is displaying fees at the section level, not at the
course. This is done with the Fees view.

Some fees apply to all sections. In TFMS those have a `section` value of
NULL. I'm converting this to the more useful 'All' string, which
duplicates what we do in the current One Stop fee search page.

In most cases when a course has a fee applied to multiple sections the
fee is exactly the same across all sections. We originally planned to
take advantage of this and just roll the duplicate fees into one fee.
But after talking to SF-IT we found that some courses do have different
fees per section. So we decided to keep the fees at the section level
and accept the duplication that comes with that choice.

With these changes the expected usage of this API has changed. Users
will probably navigate to a specific campus/term and retrieve all the
data they need from the list of courses. So the courses list view
contains everything -- subject, term and fees data. To reduce N+1
queries, getting all courses from a term now includes all
related models.