Add snapshot_builder, use snapshots for models
The performance of the views is pretty terrible, leading to very slow
response times when Jenn is pulling data. Since this data doesn't change
often, we don't need to use live views, we can use daily snapshots.

This commit

- Adds the snapshot_builder gem
- Converts the models to use the snapshot as their base table

I've deployed this to staging and it solves the performance problem.