Adjust mapping of Fall courses
[Fixes #150026375]

https://www.pivotaltracker.com/story/show/150026375

This query was written with the expectation that the `fiscal_year.year`
value would easily convert to `strm` values. But, thanks to how fiscal
years work that's not really the case.

Example

A Course Fee Occurrence has the Fiscal Year 2017 and applies to courses
in the Fall, Spring and Summer term.

The old query would convert these to strm values of

- 1173
- 1175
- 1179

But it should be

- 1173
- 1175
- 1169 <-- NOTE! This is the different one

Fiscal Year 2017 actually applies to fall term courses of the *previous*
year. You can think of Fiscal Year 2017 as beginning on September 2016
if that helps.

So, when creating the strm values here we need to handle Fall courses
differently. The year is actually 1 less than the `fiscal_year.year`
value.