Begin process of updating Ruby and Rails
Part of #33 and #34

I wasn't able to get the gems to build in my local Ruby 2.3.4, so I'm
starting the process of dockerizing this app.

What I have here is the bare minimum that will allow me to install the
gems in a container and run the tests.

This only updates the last 4.x.x.x version of Rails as a place to start.

**Why Ruby 2.5.9?**

This version of Rails still uses `BigDecimal`. I had to start with a
version of Ruby that still supported that. Once we upgrade Rails, we can
update Ruby to a newer version.

**What's up with the empty manifest.json file?**

[The newer version of Sprockets requires one](https://github.com/rails/sprockets/blob/070fc01947c111d35bb4c836e9bb71962a8e0595/UPGRADING.md#manifestjs).

This app has no JavaScript, so there's no need to put anything in that
file. I'm just making Rails happy.

**Why no build instructions?**

We can standardize this more once more upgrades are in place. For now I
run:

* `gem install lastpassify` locally (not in the container)
* `lastpassify config/database.example.yml config/database.yml`
* `docker build -t course_fees .` to build the container
* `docker run -it -v ${PWD}:/app -w /app course_fees /bin/bash` to get
  into the new container
* `bundle exec rspec` to run the tests once I'm there

The app can connect to Oracle and the tests all pass.