Update to Rails 5.2.6
Closes #34

The `peoplesoft_models` gem doesn't yet support Rails 6, so this is as
far as I can take the Rails upgrade for now.

In order to get `script/setup` to work, I had to remove
`bundle --gemfile=./vendor/integration_site/Gemfile`. We don't have a
`vendor` directory in the project, so this is vestigial.