Flatten Course relationships
This technique comes from jsonapi-resources docs

https://jsonapi-resources.com/v0.10/guide/resources.html#Flattening-a-Rails-relationship

In an older version of jsonapi resources, this code in a resource file

```
attributes :term
```

Would result in json like this

```json
{
 "attributes": {
   "term": {
     id: [term id],
     strm: [term strm],
    }
  }
}
```

But that changed and now we get

```json
{
 "attributes": {
   "term": "<Term::ActiveRecord....>"
  }
}
```

Like it's calling `to_s` on the term object and returning that.

You could convert them to relationships, changing the resource file like
so

```
belongs_to :term
```

But then the term data is no longer in the "attributes" section and our
Drupal migration fails because it expects term data to be in
"attributes"

But this "flatten the relationship" approach that the jsonapi resources
docs restores the old behavior (I think, see below). After trying it in
production, a course's fees are now displayed as

```json
{
	"data": {
		"id": "26607733",
		"type": "courses",
		"links": {
			"self": "http://course-fees.umn.edu/courses/26607733"
		},
		"attributes": {
			"class-name": "Med Term",
			"catalog-number": "1010",
			"subject": {
				"id": 383024984,
				"term_id": 176259599,
				"abbreviation": "HSM",
				"name": "Health Systems Management"
			},
			"fees": [{
				"id": 3961274619,
				"course_id": 26607733,
				"amount": "50.0",
				"fee_type": "Per Credit",
				"fee_description": "E-Learn",
				"section": "E90"
			}],
			"term": {
				"id": 176259599,
				"campus_id": "UMNCR",
				"strm": "1203",
				"name": "Spring 2020",
				"begin_date": "2020-01-13",
				"end_date": "2020-05-08",
				"current_term": "false"
			}
		}
	}
}
```

Looks right to me.

The `attributes` method is provided by ActiveRecord and takes any
ActiveRecord object and returns it as a hash of its attributes and their
values. Which is exactly what we want here.

Is it right?

We don't have a test that covers the expected json structure. Nor
is it well documented anywhere. I'm working from the drupal migration
code

https://github.umn.edu/asrweb/drupal-onestop-sites/blob/f12393e8f286979a0571565fbb038c90ef07d512/modules/custom/course_fees/includes/base.inc#L150

And from what I can see the JSON above should match what the migration
wants. But we'll want to pay close attention when we next migrate.