Test Course json structure
Closes #55

The reason this repo didn't have tests like this is because its Dev and
Test databases are empty. They have the mviews used by the models, but
those mviews contain no data.

If we were writing this app now we'd probably use something like the
Remote Record pattern to give us testable data. That is a good long-term
idea and I'm going to make an issue for that.

But, short term, in order to get this behavior under test, I'm pointing
the Test environment to the Staging database. That database contains all
the data we need.

This change also revealed that the `terms_spec` test was invalid,
pointing to a route that does not exist. Fixed.