Include fee ID in course_fees hash
Fixes #58 - please see that issue for fuller context.

[TDX ticket that identified the problem](`https://tdx.umn.edu/TDNext/Apps/29/Tickets/TicketDet?TicketID=638401)

Certain fees that should have been treated as separate were getting
combined in this query, thanks to the unique hash not producing unique
results.  This change includes the `fee_id` in the unique hash, ensuring
that different fees will be included separately.