Include the course fee ID in the fee ID hash
Actually fixes #58.

In #59, the course_fees view was updated so that two fees occurrences
with the same amount, and type for the same course would be separate if
they came from different fees.  However, I failed to account for having
the `fees` view also include this change.  Since `fees` also creates a
hash as an ID, if all of the same amounts, types, descriptions, etc.
apply for a fee, fees would also merge the two fees together.

This adds the course_fee_id to the ID hash for fees, thus ensuring that
different course fees will be kept separate.