Add and configure rack-cache
[Finishes #182331593](https://www.pivotaltracker.com/story/show/182331593)

*Overall description*

The course fees data only changes once a day, but since it isn't cached,
it is slow to load every time the endpoint is hit.

With this change, we are setting the cache in the
`ApplicationController` to last for an hour. I've tested on staging, and
with this commit, the `Cache-Control` header changes from:

`max-age=0, private, must-revalidate`

to

`max-age=3600, public`

Reloading the page is much, much faster with this change (at least until
the cache expires again).

*Notes about outdated documentation*

Below are some notes about tricky things I ran into while setting this
up, in case someone tries to set up `Rack::Cache` again.

There were [instructions on how to setup Rack::Cache](https://snippets.aktagon.com/snippets/302-how-to-setup-and-use-rack-cache-with-rails), but they were
a bit out of date.

- instead of `config/environment.rb`, we needed to use `application.rb`
or `config/environments/{ environment.rb }` to load this config.
   - I opted for `application.rb` since at least for now, we want the
   cache to behave the same way in all environments.

- the docs said if the use of `Rack::Cache` was successful, you could
run `rake middleware` and see an output of:
   - `use Rack::Cache, {:metastore=>"file:/var/cache/rack/meta", :entitystore=>"file:/var/cache/rack/body", :verbose=>true}`
   - But what I saw was `use Module`. [There is an issue about this](https://github.com/rack/rack-cache/issues/5).

- the docs also used the old `:key => value` syntax, which I updated to
`key: value`.