Store cache in folder app can access
[Finishes #182331593](https://www.pivotaltracker.com/story/show/182331593)

Follow up to #73, based on [the folders used in courses for caching](https://github.umn.edu/asrweb/courses/blob/master/config.ru).

We realized that course fees was only caching the results for each user,
and that cache wasn't shared. This change opens up the cache to be
shared by all users, by fixing the folder the cache is stored in.

Co-authored-by: Brittany Greaner <grea0022@umn.edu>
Co-authored-by: Ian Whitney <whit0694@umn.edu>