Build snapshots on deploy
Weirdly, nothing in our deploy script was building snapshots. But there
were existing snapshots in the databases. Maybe we removed the build
step accidentially? I'm not sure.

I wondered if the existing snapshots were being used. And they are. Most
of this app's models use the snapshots as their data source.

I've tested this deploy in Staging and confirmed that the snapshots are
now built via deploy.