Switching repo to utilize scenic view migrations.
Data views and snapshot builder were removed, along with many references to them.
Data views and snapshot builder were building a table and then a view on top of those tables to
increase the speed of the queries for the app. With scenic views i've switched this to create mviews,
a lot like snapshot builder was doing. With an mview though I don't have a need for both the table
and the mview, I'm just doing all of the query logic within the single mview creation. So there is now
only a single table [table_name]_daily for all of the views the applicaitonw as using.

During my commit I got a vulnerability scan warning for puma. It found a medium vulnurability and suggested
updating the version of Puma that we're using. I did not do this step and went around that vulnerability check
and did a pre-commit to skip it. I'm not sure where DEs responsibilities land for updating this so if you'd
like me to make these changes too lets talk over that. I don't use many of your libraries so I'm not sure
you'd want me taking that next step.

My story for this work is: https://www.pivotaltracker.com/story/show/187846694