Updating view names to follow ruby standard naming
Taking Ian's suggestion to switch the table names to plural for the views so that they match a better
naming standard in scenic and ruby. This will cut down on issues when you create any new migraiton,
since migrations will create your tables with a plural even when you tell it not to, then you have
to hand edit more items then needed. This naming will allow more standards to be followed.